requirements.txt
Django==1.5.1
South==0.8.1
Unipath==1.0
amqp==1.0.13
anyjson==0.3.3
billiard==2.7.3.32
celery==3.0.22
dj-database-url==0.2.2
dj-static==0.0.5
django-celery==3.0.21
django-celery-with-redis==3.0
django-nose==1.2
django-tastypie==0.10.0
django-toolbelt==0.0.1
gunicorn==18.0
kombu==2.5.14
mimeparse==0.1.3
newrelic==2.0.0.1
nose==1.3.0
psycopg2==2.5.1
python-dateutil==2.1
pytz==2013b
redis==2.7.6
six==1.4.1
static==0.4
